add new comp PresenceAvatar in messages table
